feat(core): Split AnchorService from AnchorValidator
This is important so that Ceramic nodes running with --disableAnchors can still validate AnchorCommits
